Add macOS sandboxing
This patch adds rudimentary macOS sandboxing using `sandbox_init`. While
deprecated, this seems to be the best API available for this usecase.

The current sandbox only allows changing the `file-read*` and
`file-write*` permissions. While this does not allow anything that
shouldn't be allowed, some things that are desired might not work.