Add seccomp-based Linux network filtering
This adds a new seccomp layer to the Linux sandbox, making it possible
to filter system calls which cannot be restricted by Landlock (yet).

Currently this permits all filesystem access, since this is handled by
landlock and provides the ability to enable or disable networking on top
of that.

There is no distinction between network egress and ingress, since this
is done through the connect and bind system calls and those are
identical for local and network sockets. It is only possible to restrict
the type of socket created, not the system calls used with a specific
socket domain. Restricting bind would otherwise also prevent from
binding local unix sockets.

Besides netorking, this also introduces a small rework of the execute
permission. Instead of having it integrated into `Read`, it is now
separated into `ReadAndExecute`.