Rename ReadAndExecute exception to ExecuteAndRead
This renames the `ReadAndExecute` exception to put the `Execute` at the
front, since users will really only use this when their goal is to
`Execute` something while the suffix is just to ensure its consequences
are understood.

This change should make it easier for developers to find this exception
and also help tab completion with doing its job.