Remove development disclaimer
There are no plans to make significant changes to the birdcage API in
the near future, so it should be safe for other people to use if
interested.

There also have not been any issues with the security of the provided
sandbox, which makes sense considering Birdcage just delegates to
existing OS sandboxing mechanisms.

Closes #26.