Add namespace support to Linux sandbox (#43)
The previous Linux sandbox would allow for access of abstract namespace
sockets and handled network sandboxing through fragile seccomp rules.

To both simplify the code and improve our sandboxing, this patch
introduces the usage of Linux namespaces to both clear the abstract
namespace and create a network namespace.

It is now recommended to lock down the sandbox in a newly created
process, since user namespaces cannot be created in multi-threaded
applications.

While this patch technically removes the necessity for our seccomp
network filter, it is still kept in place for redundancy and as an
alternative when namespace creation is blocked (i.e. inside docker).