Fix root filesystem exceptions on Linux
This fixes an issue with filesystem exceptions on Linux where allowing
access to `/` would prevent creating the directory to put the old root
with `pivot_root`.

Instead, the old root is now mounted on top of the new root, removing
the necessity to create a separate directory for it. By immediately
unmounting the root, the old root then gets discarded and the new root
is accessible.