Add PID namespace support
This patch changes the API to require spawning a new process to enable
the sandbox. This is necessary because PID namespaces on Linux only take
effect for new processes.

As a result, the new process will be created as PID 1 without access to
any other process through system calls like `kill` or the `procfs`
filesystem interface.

This fixes a gap in Birdcage's environment variable isolation where it
was still possible to read the unsandboxed environment from
`/proc/self/environ`.

Since the new process takes on the responsibility of PID 1 in the new
namespace, it will automatically be made parent for any orphaned
process. Currently these processes will remain zombies until the
sandboxed process is shut down.