Add separate PID 1 process on Linux
This patch uses the `clone` syscall to spawn a dedicated PID 1 process
responsible for supervising all child processes of the sandbox and
notifying the parent about their status.

As a result, it is no longer necessary to spawn the sandbox from a
dedicated process, since the original process is completely unafected
now.

This also properly handles killing hung children when killing the
parent.